Update logger to write multiple buses to one file
The 3 CAN buses are once again written to a single file. This is believed to be made possible due to the introduction of the hardware timestamp. This change also includes setting the hardware timestamp and bus timeout through configurable interface arguments. Error checking has been added for configuring each bus, allowing to issue an error message and exit if there's an issue configuring a bus.